#2199 removed home_url and replaced KiwixMainActivity
