+++ app/src/main/AndroidManifest.xml
@@ -34,14 +34,6 @@
 
 
     <activity
-      android:name=".main.KiwixNewNavigationActivity"
-      android:configChanges="orientation|keyboardHidden|screenSize"
-      android:label="@string/app_name"
-      android:launchMode="singleTop"
-      android:windowSoftInputMode="adjustPan">
-
-    </activity>
-    <activity
       android:name=".main.KiwixMainActivity"
       android:configChanges="orientation|keyboardHidden|screenSize"
       android:label="@string/app_name"
