+++ app/src/main/java/org/kiwix/kiwixmobile/di/components/KiwixActivityComponent.kt
@@ -27,8 +27,6 @@ import org.kiwix.kiwixmobile.intro.IntroModule
 import org.kiwix.kiwixmobile.language.LanguageActivity
 import org.kiwix.kiwixmobile.local_file_transfer.LocalFileTransferActivity
 import org.kiwix.kiwixmobile.main.KiwixMainActivity
-import org.kiwix.kiwixmobile.main.KiwixNewNavigationActivity
-import org.kiwix.kiwixmobile.main.KiwixReaderFragment
 import org.kiwix.kiwixmobile.nav.destination.library.LocalLibraryFragment
 import org.kiwix.kiwixmobile.nav.destination.library.OnlineLibraryFragment
 import org.kiwix.kiwixmobile.nav.destination.reader.ReaderFragment
@@ -58,14 +56,12 @@ interface KiwixActivityComponent {
   fun inject(localFileTransferActivity: LocalFileTransferActivity)
   fun inject(zimManageActivity: ZimManageActivity)
   fun inject(languageActivity: LanguageActivity)
-  fun inject(kiwixMainActivity: KiwixMainActivity)
   fun inject(kiwixSettingsActivity: KiwixSettingsActivity)
   fun inject(zimHostActivity: ZimHostActivity)
   fun inject(introActivity: IntroActivity)
   fun inject(kiwixSplashActivity: KiwixSplashActivity)
-  fun inject(kiwixNewNavigationActivity: KiwixNewNavigationActivity)
+  fun inject(kiwixMainActivity: KiwixMainActivity)
   fun inject(onlineLibraryFragment: OnlineLibraryFragment)
-  fun inject(kiwixMainFragment: KiwixReaderFragment)
 
   @Subcomponent.Builder
   interface Builder {
