+++ app/src/main/java/org/kiwix/kiwixmobile/main/KiwixMainActivity.kt
@@ -1,6 +1,6 @@
 /*
  * Kiwix Android
- * Copyright (c) 2019 Kiwix <android.kiwix.org>
+ * Copyright (c) 2020 Kiwix <android.kiwix.org>
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
  * the Free Software Foundation, either version 3 of the License, or
@@ -18,25 +18,74 @@
 
 package org.kiwix.kiwixmobile.main
 
+import android.content.Intent
 import android.os.Bundle
+import android.view.Menu
+import androidx.appcompat.view.ActionMode
+import androidx.navigation.NavController
+import androidx.navigation.findNavController
+import androidx.navigation.ui.AppBarConfiguration
+import androidx.navigation.ui.setupWithNavController
+import kotlinx.android.synthetic.main.activity_new_navigation.nav_view
 import org.kiwix.kiwixmobile.R
+import org.kiwix.kiwixmobile.core.base.BaseFragmentActivityExtensions
 import org.kiwix.kiwixmobile.core.di.components.CoreComponent
 import org.kiwix.kiwixmobile.core.main.CoreMainActivity
 import org.kiwix.kiwixmobile.kiwixActivityComponent
 
 class KiwixMainActivity : CoreMainActivity() {
+  private lateinit var navController: NavController
+  private lateinit var appBarConfiguration: AppBarConfiguration
+  private var actionMode: ActionMode? = null
+
   override fun injection(coreComponent: CoreComponent) {
     kiwixActivityComponent.inject(this)
   }
 
+  private val finishActionModeOnDestinationChange =
+    NavController.OnDestinationChangedListener { controller, destination, arguments ->
+      actionMode?.finish()
+    }
+
   override fun onCreate(savedInstanceState: Bundle?) {
     super.onCreate(savedInstanceState)
-    setContentView(R.layout.activity_kiwix_new_navigation)
+    setContentView(R.layout.activity_new_navigation)
 
-    if (savedInstanceState != null) {
-      return
+    navController = findNavController(R.id.nav_host_fragment)
+    navController.addOnDestinationChangedListener(finishActionModeOnDestinationChange)
+    appBarConfiguration = AppBarConfiguration(navController.graph)
+    nav_view.setupWithNavController(navController)
+  }
+
+  override fun onSupportActionModeStarted(mode: ActionMode) {
+    super.onSupportActionModeStarted(mode)
+    actionMode = mode
+  }
+
+  override fun onCreateOptionsMenu(menu: Menu): Boolean {
+    val onCreateOptionsMenu = super.onCreateOptionsMenu(menu)
+    menu.findItem(R.id.menu_new_navigation)?.isVisible = false
+    return onCreateOptionsMenu
+  }
+
+  override fun onSupportNavigateUp(): Boolean {
+    val navController = findNavController(R.id.nav_host_fragment)
+    return navController.navigateUp() ||
+      super.onSupportNavigateUp()
+  }
+
+  override fun onBackPressed() {
+    supportFragmentManager.fragments.filterIsInstance<BaseFragmentActivityExtensions>().forEach {
+      if (it.onBackPressed(this) == BaseFragmentActivityExtensions.Super.ShouldCall) {
+        super.onBackPressed()
+      }
     }
-    supportFragmentManager.beginTransaction()
-      .add(R.id.kiwix_fragment_containter, KiwixReaderFragment()).commit()
+  }
+
+  override fun onNewIntent(intent: Intent) {
+    super.onNewIntent(intent)
+    supportFragmentManager.fragments.filterIsInstance<BaseFragmentActivityExtensions>().forEach {
+      it.onNewIntent(intent, this)
+    }
   }
 }
