--- custom/src/main/assets/home.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <meta charset="UTF-8"/>
-    <title>REPLACE_menu_home</title>
-</head>
-<body>
-<h1 style="display:none;">REPLACE_menu_home</h1>
-</body>
-</html>
